requirements.txt
flask
azure-cognitiveservices-vision-customvision
matplotlib
Image